+++ settings.example.py
@@ -337,6 +337,7 @@ sensors = {
             "number": 0,
             "enabled": True,
             "pin_data": "P11",
+            #"driver": "native",
         },
     ]
 }